requirements.txt
opencv-python
cmake
dlib
face_recognition
setuptools
ffmpeg-python
paho-mqtt
